7. Modularize the code Refactor the time utility code
