+++ staff/gustavo-rojas/course-3/task-2/scheduler/kill_all.sh
