+++ staff/gustavo-rojas/course-3/task-2/scheduler/src/App.js
@@ -1,13 +1,11 @@
 import React, { useState, useEffect } from 'react';
 // import React from 'react';
 import './App.css';
-
+import { hasConflict,getCourseTerm,getCourseNumber,terms } from './utilities/times.js';
 const Banner = ({ title }) => (
   <h1>{ title }</h1>
 );
-const hasConflict = (course, selected) => (
-  selected.some(selection => courseConflict(course, selection))
-);
+
 
 const addScheduleTimes = schedule => ({
   title: schedule.title,
@@ -96,34 +94,17 @@ const CourseList = ({ courses }) => {
     </>
   );
 };
-const days = ['M', 'Tu', 'W', 'Th', 'F'];
 
-const daysOverlap = (days1, days2) => ( 
-  days.some(day => days1.includes(day) && days2.includes(day))
-);
 
-const hoursOverlap = (hours1, hours2) => (
-  Math.max(hours1.start, hours2.start) < Math.min(hours1.end, hours2.end)
-);
 
-const timeConflict = (course1, course2) => (
-  daysOverlap(course1.days, course2.days) && hoursOverlap(course1.hours, course2.hours)
-);
 
-const courseConflict = (course1, course2) => (
-  getCourseTerm(course1) === getCourseTerm(course2)
-  && timeConflict(course1, course2)
-);
 
-const terms = { F: 'Fall', W: 'Winter', S: 'Spring'};
 
-const getCourseTerm = course => (
-  terms[course.id.charAt(0)]
-);
 
-const getCourseNumber = course => (
-  course.id.slice(1, 4)
-);
+
+
+
+
 //var setSchedule =[];
 
 const App = () => {
